Sync bitbucket and GitHub

+++ plugins/modules/na_ontap_lun_map.py
@@ -2,7 +2,7 @@
 
 """ this is lun mapping module
 
- (c) 2018-2019, NetApp, Inc
+ (c) 2018-2022, NetApp, Inc
  # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 """
 
@@ -10,15 +10,8 @@ from __future__ import absolute_import, division, print_function
 __metaclass__ = type
 
 
-ANSIBLE_METADATA = {'metadata_version': '1.1',
-                    'status': ['preview'],
-                    'supported_by': 'certified'}
-
-
 DOCUMENTATION = """
-
 module: na_ontap_lun_map
-
 short_description: NetApp ONTAP LUN maps
 extends_documentation_fragment:
     - netapp.ontap.netapp.na_ontap
@@ -59,8 +52,6 @@ options:
     description:
     - LUN ID assigned for the map.
     type: str
-
-
 """
 
 EXAMPLES = """
@@ -125,14 +116,15 @@ from ansible.module_utils._text import to_native
 import ansible_collections.netapp.ontap.plugins.module_utils.netapp as netapp_utils
 import codecs
 from ansible.module_utils._text import to_text, to_bytes
+from ansible_collections.netapp.ontap.plugins.module_utils.netapp_module import NetAppModule
+
 
 HAS_NETAPP_LIB = netapp_utils.has_netapp_lib()
 
 
-class NetAppOntapLUNMap(object):
+class NetAppOntapLUNMap:
 
     def __init__(self):
-
         self.argument_spec = netapp_utils.na_ontap_host_argument_spec()
         self.argument_spec.update(dict(
             state=dict(required=False, type='str', choices=['present', 'absent'], default='present'),
@@ -149,24 +141,17 @@ class NetAppOntapLUNMap(object):
             ],
             supports_check_mode=True
         )
-
         self.result = dict(
             changed=False,
         )
 
-        p = self.module.params
-
-        # set up state variables
-        self.state = p['state']
-        self.initiator_group_name = p['initiator_group_name']
-        self.path = p['path']
-        self.vserver = p['vserver']
-        self.lun_id = p['lun_id']
+        self.na_helper = NetAppModule()
+        self.parameters = self.na_helper.set_parameters(self.module.params)
 
         if HAS_NETAPP_LIB is False:
             self.module.fail_json(msg="the python NetApp-Lib module is required")
         else:
-            self.server = netapp_utils.setup_na_ontap_zapi(module=self.module, vserver=self.vserver)
+            self.server = netapp_utils.setup_na_ontap_zapi(module=self.module, vserver=self.parameters['vserver'])
 
     def get_lun_map(self):
         """
@@ -176,7 +161,7 @@ class NetAppOntapLUNMap(object):
         :rtype: dict
         """
         lun_info = netapp_utils.zapi.NaElement('lun-map-list-info')
-        lun_info.add_new_child('path', self.path)
+        lun_info.add_new_child('path', self.parameters['path'])
         result = self.server.invoke_successfully(lun_info, True)
         return_value = None
         igroups = result.get_child_by_name('initiator-groups')
@@ -184,7 +169,7 @@ class NetAppOntapLUNMap(object):
             for igroup_info in igroups.get_children():
                 initiator_group_name = igroup_info.get_child_content('initiator-group-name')
                 lun_id = igroup_info.get_child_content('lun-id')
-                if initiator_group_name == self.initiator_group_name:
+                if initiator_group_name == self.parameters['initiator_group_name']:
                     return_value = {
                         'lun_id': lun_id
                     }
@@ -201,7 +186,7 @@ class NetAppOntapLUNMap(object):
         """
         # build the lun query
         query_details = netapp_utils.zapi.NaElement('lun-info')
-        query_details.add_new_child('path', self.path)
+        query_details.add_new_child('path', self.parameters['path'])
 
         query = netapp_utils.zapi.NaElement('query')
         query.add_child_elem(query_details)
@@ -233,48 +218,45 @@ class NetAppOntapLUNMap(object):
         """
         Create LUN map
         """
-        options = {'path': self.path, 'initiator-group': self.initiator_group_name}
-        if self.lun_id is not None:
-            options['lun-id'] = self.lun_id
+        options = {'path': self.parameters['path'], 'initiator-group': self.parameters['initiator_group_name']}
+        if self.parameters['lun_id'] is not None:
+            options['lun-id'] = self.parameters['lun_id']
         lun_map_create = netapp_utils.zapi.NaElement.create_node_with_children('lun-map', **options)
 
         try:
             self.server.invoke_successfully(lun_map_create, enable_tunneling=True)
         except netapp_utils.zapi.NaApiError as e:
             self.module.fail_json(msg="Error mapping lun %s of initiator_group_name %s: %s" %
-                                  (self.path, self.initiator_group_name, to_native(e)),
+                                  (self.parameters['path'], self.parameters['initiator_group_name'], to_native(e)),
                                   exception=traceback.format_exc())
 
     def delete_lun_map(self):
         """
         Unmap LUN map
         """
-        lun_map_delete = netapp_utils.zapi.NaElement.create_node_with_children('lun-unmap', **{'path': self.path, 'initiator-group': self.initiator_group_name})
+        lun_map_delete = netapp_utils.zapi.NaElement.create_node_with_children('lun-unmap', **{
+            'path': self.parameters['path'], 'initiator-group': self.parameters['initiator_group_name']})
 
         try:
             self.server.invoke_successfully(lun_map_delete, enable_tunneling=True)
         except netapp_utils.zapi.NaApiError as e:
             self.module.fail_json(msg="Error unmapping lun %s of initiator_group_name %s: %s" %
-                                  (self.path, self.initiator_group_name, to_native(e)),
+                                  (self.parameters['path'], self.parameters['initiator_group_name'], to_native(e)),
                                   exception=traceback.format_exc())
 
     def apply(self):
         netapp_utils.ems_log_event("na_ontap_lun_map", self.server)
         lun_details = self.get_lun()
-        lun_map_details = self.get_lun_map()
-
-        if self.state == 'present' and lun_details:
+        current = self.get_lun_map()
+        cd_action = self.na_helper.get_cd_action(current, self.parameters)
+        if self.parameters['state'] == 'present' and lun_details:
             self.result.update(lun_details)
-
-        if self.state == 'present' and not lun_map_details:
-            self.result['changed'] = True
-            if not self.module.check_mode:
+        self.result['changed'] = self.na_helper.changed
+        if self.na_helper.changed and not self.module.check_mode:
+            if cd_action == 'create':
                 self.create_lun_map()
-        elif self.state == 'absent' and lun_map_details:
-            self.result['changed'] = True
-            if not self.module.check_mode:
+            if cd_action == 'delete':
                 self.delete_lun_map()
-
         self.module.exit_json(**self.result)
